pyproject.toml
[project]
name = "cityscope"
version = "0.1.0"
description = "Real estate investment research data pipeline for US metros and cities"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Kaveh Karbasi", email = "kkarbasi@berkeley.edu" }
]
requires-python = ">=3.13"
keywords = ["real-estate", "data-pipeline", "census", "bls", "urban", "research"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial :: Investment",
]
dependencies = [
    "click>=8.3.2",
    "httpx>=0.28.1",
    "pandas>=3.0.2",
    "pydantic>=2.13.1",
    "pyyaml>=6.0.3",
    "rich>=15.0.0",
]


[project.urls]
Homepage = "https://github.com/kkarbasi/cityscope"
Repository = "https://github.com/kkarbasi/cityscope"
Issues = "https://github.com/kkarbasi/cityscope/issues"

[project.scripts]
cityscope = "cityscope.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cityscope"]

[dependency-groups]
dev = [
    "hatch>=1.16.5",
    "pytest>=9.0.3",
    "pytest-httpx>=0.36.2",
]
